refactor: decompose JobManager into focused coordinators
Break up the 4,295-line JobManager (52 methods) into focused services:

- IdentificationCoordinator (745 lines): disc scanning, DiscDB/TMDB/AI
  lookup, classification pipeline
- MatchingCoordinator (853 lines): episode matching, subtitle download,
  file readiness, DiscDB assignment, extras handling
- FinalizationCoordinator (721 lines): conflict resolution, file
  organization, review workflow, job completion
- CleanupService (118 lines): staging cleanup, timed cleanup, DiscDB export
- SimulationService (717 lines): all simulate_* methods for E2E testing

JobManager reduced to 1,166 lines of core orchestration: startup, drive
events, job creation, ripping, and coordinator wiring.

All coordinators use constructor injection for dependencies. Cross-coordinator
communication uses callbacks set after construction to avoid circular deps.
Public API unchanged — routes.py still imports from job_manager.

Part of #58 (Priority 1: Break Up JobManager)
